+++ gha/actions.ts
@@ -7,7 +7,7 @@ export const setupNodeAction =
 	"actions/setup-node@48b55a011bda9f5d6aeb4c2d9c7362e8dae4041e"; // v6.4.0
 
 export const setupPythonAction =
-	"actions/setup-python@a309ff8b426b58ec0e2a45f0f869d46889d02405"; // v6.2.0
+	"actions/setup-python@5fda3b95a4ea91299a34e894583c3862153e4b97"; // v7.0.0
 
 export const createGitHubAppTokenAction =
 	"actions/create-github-app-token@bcd2ba49218906704ab6c1aa796996da409d3eb1"; // v3.2.0
@@ -25,13 +25,13 @@ export const uploadArtifactAction =
 	"actions/upload-artifact@ea165f8d65b6e75b540449e92b4886f43607fa02"; // v4
 
 export const dockerLoginAction =
-	"docker/login-action@c94ce9fb468520275223c153574b00df6fe4bcc9"; // v3
+	"docker/login-action@dbcb813823bdd20940b903addbd779551569679f"; // v4.6.0
 
 export const dockerSetupBuildxAction =
 	"docker/setup-buildx-action@8d2750c68a42422c14e847fe6c8ac0403b4cbd6f"; // v3
 
 export const dockerSetupQemuAction =
-	"docker/setup-qemu-action@c7c53464625b32c7a7e944ae62b3e17d2b600130"; // v3
+	"docker/setup-qemu-action@96fe6ef7f33517b61c61be40b68a1882f3264fb8"; // v4.2.0
 
 export const dockerBuildPushAction =
 	"docker/build-push-action@10e90e3645eae34f1e60eeb005ba3a3d33f178e8"; // v6