+++ runner/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.0.3](https://github.com/dedalus-labs/hollywood/compare/runner-v0.0.2...runner-v0.0.3) (2026-08-21)
+
+
+### Chores
+
+* **deps:** bump actions/setup-node from 6.4.0 to 7.0.0 ([#65](https://github.com/dedalus-labs/hollywood/issues/65)) ([8ec2fda](https://github.com/dedalus-labs/hollywood/commit/8ec2fda24df6bd6a22b08a61e156839f518eb9e8))
+* **deps:** bump docker/build-push-action from 6.19.2 to 7.3.0 ([#64](https://github.com/dedalus-labs/hollywood/issues/64)) ([23f58b5](https://github.com/dedalus-labs/hollywood/commit/23f58b5a799cb5c2a9b4fe1e2a7a526518982903))
+
 ## 0.0.2 (2026-08-14)